Modified a method from datatTablePage.ts
